Add refactored decoders
